// dfp-piao-fe/src/app/shared/models/enums/section-status.enum.ts
export enum SectionStatusEnum {
  DA_COMPILARE = 'Da compilare',
  IN_COMPILAZIONE = 'In compilazione',
  COMPILATA = 'Compilata',
  IN_VALIDAZIONE = 'In validazione',
  VALIDATA = 'Validata',
  RICHIESTA_APPROVAZIONE = 'Richiesta Approvazione',
  APPROVATO = 'Approvato',
  PUBBLICATO = 'Pubblicata',
}